remove inappropriate tests imported from V1 examples
