+++ tests/tosca_2_0/attribute-definition/attributes-default-type-mismatch-inv.yaml
@@ -1,0 +1,27 @@
+tosca_definitions_version: tosca_2_0
+data_types:
+  resource_type:
+    derived_from: string
+    validation: { $valid_values: [ $value, ["interface", "memory", "disk"] ] }
+
+node_types:
+  server:
+    properties:
+      requested_cpus:
+        type: integer
+    attributes:
+      actual_cpus:
+        type: integer
+        description: Actual number of CPUs allocated to the node instance.
+        metadata:
+          added: 2026-05-11
+        # default value is a string instead of an integer, which should cause a type mismatch error
+        default: zero
+        validation: { $greater_or_equal: [ $value, 0 ] }
+      resource_counts:
+        type: map
+        description: count of each occurance of the supplied resource name
+        key_schema:
+          type: resource_type
+        entry_schema:
+          type: integer
