+++ tests/tosca_2_0/attribute-definition/attributes-no-type-inv.yaml
@@ -1,0 +1,27 @@
+tosca_definitions_version: tosca_2_0
+data_types:
+  resource_type:
+    derived_from: string
+    validation: { $valid_values: [ $value, ["interface", "memory", "disk"] ] }
+
+node_types:
+  server:
+    properties:
+      requested_cpus:
+        type: integer
+    attributes:
+      actual_cpus:
+        # invalid missing type
+        type: 
+        description: Actual number of CPUs allocated to the node instance.
+        metadata:
+          added: 2026-05-11
+        default: 0
+        validation: { $greater_or_equal: [ $value, 0 ] }
+      resource_counts:
+        type: map
+        description: count of each occurance of the supplied resource name
+        key_schema:
+          type: resource_type
+        entry_schema:
+          type: integer
