--- tests/tosca_2_0/property-definition/property-definition-refine-value-equal.yaml
@@ -1,19 +1,0 @@
-tosca_definitions_version: tosca_2_0
-
-capability_types:
-  Endpoint:
-    properties:
-      protocol: 
-        type: string
-        value: tcp
-
-  EnhancedEndpoint:
-    derived_from: Endpoint
-    properties:
-      protocol:
-        # protocol value cannot be refined once defined
-        # but in this case it's value is unchanged
-        # arguably this is valid
-        value: tcp
-
-
