--- tests/tosca_2_0/tosca-path/source-and-target.yaml
@@ -1,110 +1,0 @@
-tosca_definitions_version: tosca_2_0
-
-metadata:
-
-  template_name: Source and Target Example
-  template_author: Puccini
-
-data_types:
-  scalar-unit.size:
-    description: A scalar unit for size
-    derived_from: scalar
-    units:
-      gib: 1
-
-node_types:
-
-  Compute:
-    description: A compute node
-    requirements:
-    - local_storage:
-        relationship: LocalStorage
-        capability: Storage
-
-  WebServer:
-    description: A simple web server
-
-  Application:
-    derived_from: WebServer
-    properties:
-      url:
-        type: string
-    requirements:
-    - host:
-        relationship: HostedOn
-        capability: Host
-
-  Storage:
-    description: A storage node
-    capabilities:
-      storage:
-        type: Storage
-
-  BlockStorage:
-    derived_from: Storage
-    description: A block storage node
-    properties:
-      name:
-        type: string
-      volume_id:
-        type: string
-      size:
-        type: scalar-unit.size  
-  
-
-relationship_types:
-
-  HostedOn:
-    description: Relationship to a host
-
-  LocalStorage:
-    description: Relationship to local storage
-    properties:
-      location:
-        type: string
-      device:
-        type: string
-
-capability_types:
-
-  Storage:
-    description: A storage capability
-
-  Host:
-    description: A host capability
-
-service_template:
-
-  node_templates:
-
-    server:
-      type: Compute
-      attributes:
-        public_address: myapp.com
-      requirements:
-      - local_storage:
-          node: Storage
-          relationship:
-            properties:
-              # Within relationships you can use "SOURCE" and "TARGET" keywords
-              # For both the "get_property" and "get_attribute" functions
-              location: { concat: [ /mnt/, { get_property: [ TARGET, volume_id ] } ] }
-              device: /dev/hda1
-
-    app:
-      type: Application
-      properties:
-        # Another special kind of target is "HOST", which can be used in nodes and their capabilities
-        # It refers to the target of any "HostedOn" relationship from the node
-        url: { get_attribute: [ HOST, public_address ] }
-      requirements:
-      - host:
-          capability: Host
-          relationship: HostedOn
-
-    storage:
-      type: BlockStorage
-      properties:
-        name: storage
-        volume_id: bucket
-        size: 20 gib
